pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "This app analizes websites onto SEO suitability"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "ruff>=0.14.5",
]
